New enemy mobs
Adding some new enemy mobs from the unity asset stores that might be useful.
They are the imp, the human soldier with  armor and shield, the orc on wolf, the undead and the dragon.
The dragon may serve as flying unit while the imp(red skin) may be serve as fire-proof or anti-spell unit.